requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
ndg-httpsclient==0.5.1
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
